move observables to their own submodule (to prepare for more observables)
